+++ src/client/components/forms/parts/workData.jsx
@@ -25,7 +25,7 @@ const Input = require('react-bootstrap').Input;
 const Identifiers = require('./identifiers.jsx');
 const Select = require('../../input/select2.jsx');
 
-const validators = require('../../validators');
+const validators = require('../../../helpers/react-validators');
 
 const WorkData = React.createClass({
 	displayName: 'workDataComponent',
