+++ src/client/helpers/react-validators.js
